requirements.txt
# Core Data Science and Agent Framework
dspy-ai>=0.1.9
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
statsmodels>=0.14.0
scipy>=1.9.0

# Streamlit UI
streamlit>=1.25.0

# Visualization
plotly>=5.10.0
matplotlib>=3.6.0 # Often a dependency for statsmodels plots or advanced Plotly features

# LlamaIndex for Retrieval
llama-index-core>=0.10.0
llama-index-embeddings-openai>=0.1.0 # For OpenAI embeddings

# OpenAI (if directly used or for DSPy/LlamaIndex OpenAI integrations)
openai>=1.0.0

# Excel File Handling
openpyxl>=3.0.0

# Network Graphs (for correlation networks)
networkx>=3.0

# Flask Backend
flask>=2.2.0
flask-sqlalchemy>=3.0.0
flask-cors>=3.0.0 # For development, might be handled by gateway in production

# Environment Variable Management
python-dotenv>=1.0.0

# WSGI HTTP Server for Flask (Production)
gunicorn>=20.1.0

# Other utilities that might be implicitly used or good to have
# (Add if specific needs arise, e.g., specific database drivers if not SQLite)
# Example: psycopg2-binary (for PostgreSQL)
